ref: transform popular movies in component
